Dropdown Resize
When the dropdown is placed on top, the placement of the list is wrong once the user starts filtering with the search bar. This fix resets the placement/size of the dropdown bar as the displayed list content changes.

Note: At the moment, it works but is not optimized. There are more resizing calls made than is needed.